rf: move slurm tests to submitter tests
